split use-metabase-query.unit.spec.tsx into multiple files
